requirements.txt
Django==2.0
django-js-asset==0.1.1
django-mptt==0.9.0
django-mptt-admin==0.5.0
django-taggit==0.22.2
PyGithub==1.35
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
